requirements.txt
flask
flask_cors
requests
toolforge
chardet
rdflib
antlr4-python3-runtime==4.7.2
SPARQLWrapper
jsonasobj
PyJSG